+++ core/reporting/reporting.hpp
@@ -6,4 +6,8 @@
 #pragma once
 
 #include "report.hpp"
+#include "report_format.hpp"
+#include "report_formatter.hpp"
 #include "report_generator.hpp"
+#include "report_options.hpp"
+#include "report_section.hpp"